Fold the Responses reasoning replay into the reasoning-fidelity suites
The two model-specific suites, `gpt6-reasoning-item.test.ts` and
`test_gpt6_reasoning_item.py`, are removed. The behaviour they pinned -- a
Responses-protocol client records a reasoning item's `encrypted_content` from
`response.output_item.done` alone, and the replayed input item is rebuilt from
that value with no `id` and never the in-progress ciphertext -- now lives in the
existing parameterized suites as a second case table,
`RESPONSES_REASONING_CASES`, covering `gpt-6` (`GPT6Client`) and
`openai-responses` (`OpenaiResponsesClient`). `runTurnAndReplay` /
`_run_turn_and_replay` also hand back the whole model input so the new case can
assert on the serialized request.

The skills' `reference/data-models.md` no longer names a model in the
`thinking_level` bullet; it states the generic rule that AgentHub maps each level
to the closest effort the model supports.
