+++ changelog/unreleased/2026-09-09-gpt-6-astra.md
@@ -29,9 +29,9 @@
   was added. For reasoning items, `encrypted_content` may be incomplete while the item is in
   progress. Use the reasoning item from the corresponding `response.output_item.done` event
   when passing it as input to a subsequent request." The client comment now cites that
-  sentence, and a paired offline test pins the rule by streaming a truncated
-  `encrypted_content` on the added event and asserting only the completed value reaches the
-  replayed reasoning item.
+  sentence, and the reasoning-fidelity suites in both languages gained a Responses-protocol
+  case table that pins the rule by streaming a truncated `encrypted_content` on the added
+  event and asserting only the completed value reaches the replayed reasoning item.
 - `llmsdk_docs/gpt6/` snapshots the model page, the GPT-6 model guide, the reasoning guide and
   the prompt-caching guide; the `POST /v1/responses` reference snapshot is refreshed, and
   `llmsdk_docs/openai_responses/docs/openai-responses-streaming-events.md` adds a labelled
