+++ changelog/unreleased/2026-09-09-gpt-6-astra.zh.md
@@ -27,8 +27,9 @@
   参考文档对 `response.output_item.added` 的说明是："The output item that was added. For
   reasoning items, `encrypted_content` may be incomplete while the item is in progress. Use the
   reasoning item from the corresponding `response.output_item.done` event when passing it as
-  input to a subsequent request."。客户端注释现引用该句，并新增一对离线测试：在 added 事件上
-  下发被截断的 `encrypted_content`，断言只有完成事件的取值进入回放的 reasoning item。
+  input to a subsequent request."。客户端注释现引用该句，两种语言的 reasoning fidelity 测试套件
+  也新增了一张 Responses 协议用例表：在 added 事件上下发被截断的 `encrypted_content`，断言只有
+  完成事件的取值进入回放的 reasoning item。
 - `llmsdk_docs/gpt6/` 收录模型页、GPT-6 模型指南、推理指南与提示缓存指南；`POST /v1/responses`
   参考快照一并刷新，`llmsdk_docs/openai_responses/docs/openai-responses-streaming-events.md`
   新增 streaming-events 参考文档的标注摘录，覆盖 output item 生命周期事件、`response.reasoning_*`
