+++ skills/agenthub-typescript/reference/data-models.md
@@ -37,7 +37,7 @@ Fields:
 - `temperature` (`number`): Sampling temperature; support varies by model.
 - `tools` (`ToolSchema[]`): Tools with `name`, `description`, and optional JSON Schema `parameters`.
 - `thinking_summary` (`boolean`): Request a thinking summary when supported. Gemini returns summaries only on some models (3.1-pro does; the 3.5-3.8 flash models do not).
-- `thinking_level` (`ThinkingLevel`): `NONE`, `LOW`, `MEDIUM`, `HIGH`, `XHIGH`, or `MAX`. GPT-6 rejects `none`, so the client sends `low` for `NONE`.
+- `thinking_level` (`ThinkingLevel`): `NONE`, `LOW`, `MEDIUM`, `HIGH`, `XHIGH`, or `MAX`. AgentHub maps each level to the closest effort the model supports, so any level is safe to pass.
 - `tool_choice` (`ToolChoice`): `auto`, `required`, `none`, or a list of tool names; support varies by model.
 - `system_prompt` (`string`): System instruction text.
 - `prompt_caching` (`PromptCaching`): `ENABLE`, `DISABLE`, or `ENHANCE`.
